getting closer
so I've got the pieces and spawning mostly normal. in order to detect individual blocks I'm going to give each individual block it's own area. this will be used to detect tetrist in the future, but for now i am focused on getting everything to work perfectly in a pixel perfect way